+++ src/main/java/walkingkooka/spreadsheet/server/parser/SpreadsheetParserSelectorEditContextTesting.java
@@ -20,7 +20,7 @@ package walkingkooka.spreadsheet.server.parser;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterContextTesting;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterProviderTesting;
 import walkingkooka.spreadsheet.parser.SpreadsheetParserContextTesting;
-import walkingkooka.spreadsheet.parser.SpreadsheetParserProviderTesting;
+import walkingkooka.spreadsheet.parser.provider.SpreadsheetParserProviderTesting;
 
 public interface SpreadsheetParserSelectorEditContextTesting<C extends SpreadsheetParserSelectorEditContext> extends SpreadsheetParserProviderTesting<C>,
     SpreadsheetParserContextTesting<C>,
